Contact Mail Fixed
Tried to add direct email sender using Forms, but it opens the default mail application on the device using the values I gave in form tag.
That works for now. In the near future I will add a php script to help me develop this feature. Don't know how thou.